linear regression example with graph written with Python API v1.7.0
