+++ examples/src/main/scala/org/platanios/tensorflow/examples/python2scala/checkpoint
@@ -1,0 +1,3 @@
+model_checkpoint_path: "linear-regression-pull-request"
+all_model_checkpoint_paths: "virgin-linear-regression-pull-request"
+all_model_checkpoint_paths: "linear-regression-pull-request"
